backend/pyproject.toml
[project]
name = "narrativeshield"
version = "1.0.0"
description = "AI-Powered Disinformation Detection Platform — SciComp v2.0"
authors = [
    { name = "Ajay Kumar Reddy", email = "ajaykumarreddykrishnareddygari@gmail.com" },
    { name = "Thanvarshini VR",  email = "thanu2562005@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11"

# FIX: Moved ALL required packages into the main dependencies block.
# Now a simple `uv sync` installs everything you need automatically.
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "google-generativeai>=0.5.4",
    "textstat>=0.7.3",
    "python-dotenv>=1.0.1",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

# ── uv settings ───────────────────────────────────────────────────────────────
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",       
    "ruff>=0.4.0",         
]

# ── ruff linter (optional but nice) ──────────────────────────────────────────
[tool.ruff]
line-length    = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]